requirements.txt
python-telegram-bot==20.7
flask==2.3.2
pyqrcode==1.2.1
pypng==0.0.20
pillow==10.4.0
gunicorn==23.0.0
psycopg2-binary==2.9.10